Unify use of the rid parameter of msgRecv, msgRespond functions

+++ dma/imx6ull-sdma/imx6ull-sdma.c
@@ -735,7 +735,7 @@ static int dev_ctl(msg_t *msg)
 static void worker_thread(void *arg)
 {
 	msg_t msg;
-	unsigned rid;
+	unsigned long rid;
 
 	(void)arg;
